Moved slack rendering code into backends/slack.py
